feat(Action): add displayMode support (#823)
