+++ packages/components/src/Actions/index.js
@@ -1,6 +1,7 @@
 import Action from './Action';
+import ActionButton from './ActionButton';
 import ActionDropdown from './ActionDropdown';
 import ActionSplitDropdown from './ActionSplitDropdown';
 import Actions from './Actions.component';
 
-export { Action, ActionDropdown, ActionSplitDropdown, Actions };
+export { Action, ActionButton, ActionDropdown, ActionSplitDropdown, Actions };
